DockerFile/filter/requirements.txt
biopython==1.79
plotly>=5.18.0
numpy
